Pull out camelcase stuff into utils
